Add renderer for basic structure- the major cols
Here we use vanilla javascript and divs. D3 provides methods to build
axes with SVGs but not divs. It may be that this could be refactored to
use SVG, but since I've already picked up 2 APIs and don't want to pick
up a 3rd (and because it was acting wonky on my browser), I decided to
let it go.