Change ID/class generation to not break D3:
So apparently D3 will attach __data__ to every object appended. That's
fine. When we re-render, it knows which objects=which data and it
wont create redudant objects because it sees data hasn't changed.

To do this, we use a key function. I'd like to generate a key based on
the major row and the filter name, but since they're both full of
arbitrary characters, and since apparently no escape function is
suitable for D3, I use the index instead, prefixed by a character (ie
"i" + index)

We're right now doing this for class and data-object key.